added test files, battle file, and cleaned character files
