+++ Character_Helper_Functions.py
@@ -11,10 +11,9 @@ def set_attributes(value, attribute):
         int(value)
         return int(value)
     except ValueError:
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
-        print("Value must be an integer.", attribute.upper(), " set to 0")
-        print("Use", attribute.upper(), "setter function to change attribute")
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+
+        error_message("Value must be an integer. " + attribute.upper() +
+                      " set to 0. Use " + attribute.upper() + " setter function to change attribute")
         return 0
 
 
@@ -36,9 +35,7 @@ def check_enum(enumerator, value):
     if is_valid:
         return value
     else:
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
-        print("Value must be from enum ", enumerator, " Value has been set to N/A")
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+        error_message("Value must be from enum " + enumerator + " Value has been set to N/A")
         return "na"
 
 
@@ -46,7 +43,14 @@ def check_boolean(value):
     if value == True or value == False:
         return value
     else:
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
-        print("Value must be TRUE or FALSE. Value has been set to FALSE. USE SETTER TO CHANGE")
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+        error_message("Value must be TRUE or FALSE. Value has been set to FALSE. USE SETTER TO CHANGE")
         return False
+
+
+def error_message(message):
+    size = len(message)
+    print()
+    print("!" * size)
+    print(message)
+    print("!" * size)
+    print()
